Remove images by tag (not by id).
Removing by id doesn't work if an image has more than one tag. This
change retrieves the tags for all remove images and passes them to
docker rmi instead of using the ids directly.

The action above is hard to achieve with Docker 1.9 so this patch
includes a hack to upgrade Docker to 1.10. This can be removed when
that version shows up in Alpine stable.